requirements.txt
# Core AI/ML
pyautogen>=0.2.0,<0.3.0
openai>=1.0.0
anthropic>=0.18.0
tiktoken>=0.5.0

# Web Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
websockets>=12.0
python-multipart>=0.0.6
pydantic>=2.5.0

# CLI & Terminal
colorama>=0.4.6
termcolor>=2.3.0
rich>=13.7.0

# Configuration
python-dotenv>=1.0.0
pyyaml>=6.0.1

# HTTP Clients
requests>=2.31.0
httpx>=0.25.0
aiohttp>=3.9.0

# DevOps Integrations (optional - install as needed)
kubernetes>=28.1.0
# boto3>=1.33.0           # AWS SDK
# azure-identity>=1.15.0  # Azure SDK
# google-cloud>=0.34.0    # GCP SDK

# Database Clients (optional)
# redis>=5.0.0
# pika>=1.3.0             # RabbitMQ
# neo4j>=5.14.0
# pymysql>=1.1.0
# psycopg2-binary>=2.9.9  # PostgreSQL

# Email (optional)
# sendgrid>=6.11.0

# Monitoring Clients (optional)
# prometheus-client>=0.19.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0

# Code Quality
# pylint>=3.0.0
# black>=23.12.0
# isort>=5.13.0
